+++ metadata.rb
@@ -6,7 +6,7 @@ source_url 'https://github.com/IshentRas/cookbook-openshift3'
 issues_url 'https://github.com/IshentRas/cookbook-openshift3/issues'
 description 'Installs/Configures Openshift 3.x (>= 3.1)'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '1.10.17'
+version '1.10.18'
 supports 'redhat', '>= 7.1'
 supports 'centos', '>= 7.1'
 
@@ -23,3 +23,4 @@ recipe 'cookbook-openshift3::master_config_post', 'Configure Post actions for ma
 recipe 'cookbook-openshift3::nodes_certificates', 'Configure certificates for nodes'
 recipe 'cookbook-openshift3::node', 'Configure node'
 recipe 'cookbook-openshift3::etcd_cluster', 'Configure HA ETCD cluster'
+recipe 'cookbook-openshift3::adhoc_uninstall', 'Adhoc action for uninstalling Openshit from server'