+++ resources/openshift_delete_host.rb
@@ -1,0 +1,12 @@
+#
+# Cookbook Name:: cookbook-openshift3
+# Resources:: openshift_delete_host
+#
+# Copyright (c) 2015 The Authors, All Rights Reserved.
+
+provides :openshift_delete_host
+resource_name :openshift_delete_host
+
+actions :delete
+
+default_action :delete